chore: factor out a showPopup function to handle rendering and positioning
